Flatten util package structure
