+++ transform/rotate_test.go
@@ -5,7 +5,6 @@ import (
 	"testing"
 
 	"github.com/anthonynsimon/bild/util"
-	"github.com/anthonynsimon/bild/util/compare"
 )
 
 func TestRotate(t *testing.T) {
@@ -172,7 +171,7 @@ func TestRotate(t *testing.T) {
 
 	for _, c := range cases {
 		actual := Rotate(c.value, c.angle, c.options)
-		if !compare.RGBAImageEqual(actual, c.expected) {
+		if !util.RGBAImageEqual(actual, c.expected) {
 			t.Errorf("%s: expected: %#v, actual: %#v", "Rotate "+c.description, util.RGBAToString(c.expected), util.RGBAToString(actual))
 		}
 	}
@@ -243,7 +242,7 @@ func TestFlipH(t *testing.T) {
 
 	for _, c := range cases {
 		actual := FlipH(c.value)
-		if !compare.RGBAImageEqual(actual, c.expected) {
+		if !util.RGBAImageEqual(actual, c.expected) {
 			t.Errorf("%s: expected: %#v, actual: %#v", "FlipH", util.RGBAToString(c.expected), util.RGBAToString(actual))
 		}
 	}
@@ -314,7 +313,7 @@ func TestFlipV(t *testing.T) {
 
 	for _, c := range cases {
 		actual := FlipV(c.value)
-		if !compare.RGBAImageEqual(actual, c.expected) {
+		if !util.RGBAImageEqual(actual, c.expected) {
 			t.Errorf("%s: expected: %#v, actual: %#v", "FlipV", util.RGBAToString(c.expected), util.RGBAToString(actual))
 		}
 	}
